# packages/shared-types/src/types/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Pinaka API",
    "version": "1.0.0",
    "description": "API documentation generated from canonical schema registry"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
    "/api/v1/properties": {
      "get": {
        "summary": "List Propertys",
        "tags": [
          "Property"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 1000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of Propertys",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Property",
        "tags": [
          "Property"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PropertyCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Property created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/properties/{id}": {
      "get": {
        "summary": "Get Property by ID",
        "tags": [
          "Property"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Property details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Property",
        "tags": [
          "Property"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PropertyUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Property updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Property",
        "tags": [
          "Property"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Property deleted"
          }
        }
      }
    },
...
    "/api/v1/notifications": {
      "get": {
        "summary": "List Notifications",
        "tags": [
          "Notification"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 1000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of Notifications",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Notification",
        "tags": [
          "Notification"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Notification created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/{id}": {
      "get": {
        "summary": "Get Notification by ID",
        "tags": [
          "Notification"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notification details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Notification",
        "tags": [
          "Notification"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Notification updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Notification",
        "tags": [
          "Notification"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Notification deleted"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "PropertyCreate": {
        "type": "object",
        "description": "Schema for creating a Property"
      },
      "PropertyUpdate": {
        "type": "object",
        "description": "Schema for updating a Property"
      },
      "PropertyResponse": {
        "type": "object",
        "description": "Response schema for Property"
      },
...
      "NotificationCreate": {
        "type": "object",
        "description": "Schema for creating a Notification"
      },
      "NotificationUpdate": {
        "type": "object",
        "description": "Schema for updating a Notification"
      },
      "NotificationResponse": {
        "type": "object",
        "description": "Response schema for Notification"
      },
...
    }
  },
...
}